# database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package database

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Task struct {
	ID     int64
	Name   string
	Info   string
	Isdone bool
}

type User struct {
	ID         int32
	EmailToken string
	JwtToken   string
	Name       string
	CreatedAt  pgtype.Timestamp
	UpdatedAt  pgtype.Timestamp
}