# docs/swagger.yaml
basePath: /
definitions:
  database.AddTaskParams:
    properties:
      info:
        type: string
      name:
        type: string
    type: object
  database.Task:
    properties:
      id:
        type: integer
      info:
        type: string
      isdone:
        type: boolean
      name:
        type: string
    type: object
  database.UpdateTaskParams:
    properties:
      id:
        type: integer
      info:
        type: string
      isdone:
        type: boolean
      name:
        type: string
    type: object
  structures.JWTToken:
    properties:
      accessToken:
        type: string
    type: object
  structures.User:
    properties:
      email:
        type: string
      family_name:
        type: string
      given_name:
        type: string
      id:
        type: string
      locale:
        type: string
      name:
        type: string
      picture:
        type: string
      verified_email:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: parampreets.rai@thewitslab.com
    name: Parampreet Singh Rai
  description: API for CRUD Operations or a TODO List.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CRUD GO API
  version: "1.0"
paths:
  /:
    get:
      description: get list of all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Task'
        "400":
          description: Bad Request
          schema: {}
      summary: Show all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Add new task to ToDo List as not done
      parameters:
      - description: Task object
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/database.AddTaskParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Task'
        "400":
          description: Bad Request
          schema: {}
      summary: Adds a task
      tags:
      - tasks
  /{id}:
    delete:
      description: Delete a task with given id
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema: {}
      summary: Delete a task
      tags:
      - tasks
  /edit:
    post:
      consumes:
      - application/json
      description: Edit the task with the given id
      parameters:
      - description: Update Task Params
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/database.UpdateTaskParams'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema: {}
      summary: Edit a task
      tags:
      - tasks
  /login:
    get:
      consumes:
      - application/json
      description: Redirects the user to Google's OAuth 2.0 authentication page to
        start the login process.
      produces:
      - application/json
      responses:
        "302":
          description: Redirected to Google OAuth
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema: {}
      summary: Redirect to Google OAuth
      tags:
      - Authentication
  /profile:
    post:
      consumes:
      - application/json
      description: Get profile data with JWT Token
      parameters:
      - description: JWT Token
        in: body
        name: jwt
        required: true
        schema:
          $ref: '#/definitions/structures.JWTToken'
      produces:
      - application/json
      responses:
        "302":
          description: User details
          schema:
            $ref: '#/definitions/structures.User'
        "401":
          description: Unauthorized
      summary: Get the profile data without password
      tags:
      - JWT
swagger: "2.0"